♻️ refactor: extract UserSection component on header
