+++ src/components/header/UserSection.module.css
@@ -1,0 +1,30 @@
+.userInfoContainer {
+  position: relative;
+  display: flex;
+  align-items: center;
+}
+
+.userInfoContainer::before {
+  content: '';
+  border: 0.5px solid var(--gray-300);
+  height: 34px;
+}
+
+.userInfoButton.userInfoButton {
+  margin-left: 16px;
+  display: flex;
+  align-items: center;
+  background-color: transparent;
+}
+
+.myMenu {
+  position: absolute;
+  top: 75px;
+  right: 0;
+}
+
+@media screen and (min-width: 768px) {
+  .myMenu {
+    top: 60px;
+  }
+}
